+++ src/server/types/kras-request.ts
@@ -4,11 +4,17 @@ export interface KrasRequestQuery {
   [key: string]: string;
 }
 
-export interface KrasRequest {
+export interface BasicKrasRequest {
   url: string;
   target: string;
   query: KrasRequestQuery;
   method: string;
   headers: IncomingHttpHeaders;
   content: string;
+}
+
+export interface KrasRequest extends BasicKrasRequest {
+  encrypted: boolean;
+  remoteAddress: string;
+  port: string;
 }